Reorganizes CUnit test.main.c into separately compilable test suites
